requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==21.4.0
certifi==2022.6.15
cffi==1.15.1
cryptography==37.0.4
dj-database-url==1.0.0
Django==4.1
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
environ==1.0
gunicorn==20.1.0
h11==0.13.0
idna==3.3
mysqlclient==2.1.1
numpy==1.23.1
outcome==1.2.0
pandas==1.4.3
pip==22.2.2
psycopg2-binary==2.9.5
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
selenium==4.3.0
setuptools==63.4.1
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.1
trio==0.21.0
trio-websocket==0.9.2
tzdata==2021.1
urllib3==1.26.11
wheel==0.37.1
whitenoise==6.2.0
wincertstore==0.2
wsproto==1.1.0
